+++ packages/layout/src/components/SideNav/MedlySidenavHeader/MedlySidenavHeader.test.tsx
@@ -1,4 +1,4 @@
-import { cleanup, render } from '@test-utils';
+import { cleanup, render, screen } from '@test-utils';
 import React from 'react';
 import SidenavContext from '../SideNav.context';
 import { MedlySidenavHeader } from './MedlySidenavHeader';
@@ -26,5 +26,20 @@ describe('Medly sidenav header', () => {
             </SidenavContext.Provider>
         );
         expect(container).toMatchSnapshot();
+    });
+
+    it('should render properly even when custom company name and logo are passed', () => {
+        const DummyLogo = () => <span>My Logo</span>;
+        const DummyName = () => <span>My Name</span>;
+
+        render(
+            <SidenavContext.Provider
+                value={{ activeItem: 'dummy', isHovered: true, isExpanded: true, activeItemChangeHandler: () => null }}
+            >
+                <MedlySidenavHeader companyLogo={DummyLogo} companyName={DummyName} />
+            </SidenavContext.Provider>
+        );
+        expect(screen.getByText('My Logo')).toBeInTheDocument();
+        expect(screen.getByText('My Name')).toBeInTheDocument();
     });
 });